@skottmckay 2024-12-23
Are there any unit tests for the new code? We should be able to test that the expected cache files are created in the right places and that things like invalid cache key values are rejected.